geo-audit-agencies/requirements.txt
openai>=1.30.0
pyyaml>=6.0
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.65.0
tenacity>=8.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
numpy>=1.26.0
tiktoken>=0.7.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.12.0
